Updated course structure and added chapter 1.1
To maintain shorter notebooks, the course units will be split into more
selctions. One section for each chapter. The solutions will also be
separated from the excercises to make practice more self-learning.

This commit added the first part of the first chapter in the first
course unit. It is an expansion of the introduction to the CLI and
focusses a bit more on a few useful commands for beginners.